server/requirements.txt
kafka-python
pandas==1.3
scikit-learn
seaborn
datetime
streamlit